refactor: change router
