+++ src/router/routes.tsx
@@ -1,0 +1,26 @@
+import React, { ReactNode } from 'react'
+import About from '../pages/About'
+import NotFound from '../pages/NotFound'
+import { Navigate } from 'react-router-dom'
+import Main from '@/pages/Main'
+import FormPage from '@/pages/FormPage'
+
+export enum ERoutes {
+  About = 'about',
+  Main = 'main',
+  Root = '/',
+  Form = 'form',
+}
+
+type Route = {
+  path: string
+  element: ReactNode
+}
+
+export const routes: Route[] = [
+  { path: ERoutes.About, element: <About /> },
+  { path: ERoutes.Main, element: <Main /> },
+  { path: ERoutes.Form, element: <FormPage /> },
+  { path: ERoutes.Root, element: <Navigate to={ERoutes.Form} replace={true} /> },
+  { path: '*', element: <NotFound /> },
+]
