+++ src/types/index.ts
@@ -1,12 +1,5 @@
 import type { FormData } from '@components/MyForm/MyForm'
 
-export enum ERoutes {
-  About = 'about',
-  Main = 'main',
-  Root = '/',
-  Form = 'form',
-}
-
 export const genders = ['male', 'female', 'other']
 export type Genders = (typeof genders)[number]
